# apps/server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "operationId": "ProjectsController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "项目实体数组",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 100
                          },
                          "description": {
                            "type": "string",
                            "maxLength": 500
                          },
                          "agentIds": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "sessionIds": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "skillIds": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "fileIds": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "id": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "updatedAt": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          }
                        },
                        "required": [
                          "name",
                          "agentIds",
                          "sessionIds",
                          "skillIds",
                          "fileIds",
                          "id",
                          "createdAt",
                          "updatedAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token（UNAUTHORIZED）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "机器可读错误码（SCREAMING_SNAKE_CASE，如 NOT_FOUND / VALIDATION_ERROR）"
                    },
                    "message": {
                      "type": "string",
                      "description": "面向用户的可读文案"
                    },
                    "details": {
                      "description": "结构化补充信息（如 VALIDATION_ERROR 的字段级 issue 列表）"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误（INTERNAL_ERROR）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "机器可读错误码（SCREAMING_SNAKE_CASE，如 NOT_FOUND / VALIDATION_ERROR）"
                    },
                    "message": {
                      "type": "string",
                      "description": "面向用户的可读文案"
                    },
                    "details": {
                      "description": "结构化补充信息（如 VALIDATION_ERROR 的字段级 issue 列表）"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "项目列表（按更新时间倒序）",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "ProjectsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "agentIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "sessionIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "skillIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "fileIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "新创建的项目",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 100
                        },
                        "description": {
                          "type": "string",
                          "maxLength": 500
                        },
                        "agentIds": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "sessionIds": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "skillIds": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "fileIds": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "id": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "updatedAt": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "name",
                        "agentIds",
                        "sessionIds",
                        "skillIds",
                        "fileIds",
                        "id",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数校验失败（VALIDATION_ERROR，details 含字段级 issue）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "机器可读错误码（SCREAMING_SNAKE_CASE，如 NOT_FOUND / VALIDATION_ERROR）"
                    },
                    "message": {
                      "type": "string",
                      "description": "面向用户的可读文案"
                    },
                    "details": {
                      "description": "结构化补充信息（如 VALIDATION_ERROR 的字段级 issue 列表）"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token（UNAUTHORIZED）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "机器可读错误码（SCREAMING_SNAKE_CASE，如 NOT_FOUND / VALIDATION_ERROR）"
                    },
                    "message": {
                      "type": "string",
                      "description": "面向用户的可读文案"
                    },
                    "details": {
                      "description": "结构化补充信息（如 VALIDATION_ERROR 的字段级 issue 列表）"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误（INTERNAL_ERROR）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "机器可读错误码（SCREAMING_SNAKE_CASE，如 NOT_FOUND / VALIDATION_ERROR）"
                    },
                    "message": {
                      "type": "string",
                      "description": "面向用户的可读文案"
                    },
                    "details": {
                      "description": "结构化补充信息（如 VALIDATION_ERROR 的字段级 issue 列表）"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建项目",
        "tags": [
          "projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "Growth OS API",
    "description": "Growth OS 后端服务 API 文档（自动生成）。\n\n**响应信封**：成功响应统一为 `{ data: T }`（204 除外）；失败统一为 `ApiErrorEnvelope { code, message, details? }`，`code` 为机器可读错误码，前端按码分支处理。\n\n**鉴权**：除 health 域外，所有端点均需 Bearer JWT（Supabase Auth access token）——点击右上角 Authorize 输入 token。",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}